Database Set-up, Version 1
The files added in this commit are all involved in the initial database setup through mySQL connector.

login_info.py stores the details needed for database_setup.py to succesfully log in to the mySQL database.

database_setup.py connects to the database, creates the database for this application if it does not already exist, and then creates the tables to be used in the database.