Database Setup v0.9
All parts of database setup are working except for one. The modified dataframe of the credit data is causing issues, and will be fixed later.

However, all other datasets are properly loaded from their source, modified if necessary, and then uploaded to the relevant database. This includes the additional loan dataset that is from an API source instead of a file source.