Database Setup v1.0
With this version, database setup is fully functional after fixing several issues with the credit dataframe.

While there are still some improvements I'd like to make, they need to wait until the main input/output loop of the program is implemented.

I have started the file for the I/O section as menu.py. However, it will be combined into the cdw_sapp.py file as the main function, once it is ready for testing.