Modified function that updates customer info.
Changed the aforementioned function to use the customer's social security number as a parameter instead of the credit card number, because the SSN is the primary key.

also added functionality to allow the user to change any part of the customer's details except for the SSN and credit card number.